공지사항 작성 뷰 수정

+++ src/main/java/com/mycompany/webapp/email/service/EmailService.java
@@ -474,8 +474,10 @@ public class EmailService implements IEmailService {
 		log.info("실행");
 		int row = 0;
 		List<Integer> contentIdList = emailRepository.selectIsDeletedEmail();
-		for(int contentId : contentIdList) {
-			row = emailRepository.deleteEmailContent(contentId);
+		if(contentIdList.size() != 0) {
+			for(int contentId : contentIdList) {
+				row = emailRepository.deleteEmailContent(contentId);
+			}
 		}
 		return row;
 	}
